# openapi.yaml
openapi: 3.0.3
info:
  title: Cohort — Coding Interview Session API
  version: "1.0.0"
  description: |
    Backend contract inferred from the frontend's service layer
    (`frontend/src/services/types.ts` and its mock implementations under
    `frontend/src/services/mock/`). The frontend currently talks to mocks only;
    this document describes the real backend those mocks stand in for.

    **Authentication:** none of these endpoints require authentication. There
    are no accounts, logins, or API keys anywhere in the frontend (see
    `docs/specs.md` §2 Non-Goals). The session ID itself acts as the sole
    authorization token — anyone who has it can read and join that session
    (`docs/specs.md` §8).

    **Not covered by REST:** real-time document sync and presence (cursors,
    participant list, live output sharing) happen over a separate WebSocket
    channel, opened via `CollabApi.connect(sessionId)`
    (`frontend/src/services/mock/collabApi.ts`). That interface passes
    caller-defined messages as opaque payloads (`send(message: unknown)`), so
    the frontend contract does not mandate a specific wire format — only that
    a realtime channel exists per session. It is documented below as a single
    path for completeness, since OpenAPI has no native WebSocket support.

    **Code execution:** unlike the general-purpose editor, interview problems
    are graded like LeetCode — a candidate's Python solution is checked
...
security: []
paths:
...
  /api/problems/{id}/run:
    post:
      operationId: runProblem
      summary: Run a candidate's solution against a problem's visible test cases
      description: |
        Maps to `ProblemsApi.runTests(problemId, code)`. Executes server-side
        in a sandboxed subprocess (`backend/src/backend/executor.py`) with a
        timeout; the candidate's code and the visible test cases are the
        only inputs, and per-test input/expected/actual values are returned
        so the UI can show exactly what ran.
      parameters:
        - $ref: "#/components/parameters/ProblemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: |
            Grading finished (this does not imply the tests passed — check
            `allPassed` / `error`).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunResponse"
        "404":
          description: No problem with this ID.
  /api/problems/{id}/submit:
    post:
      operationId: submitProblem
      summary: Run a candidate's solution against visible and hidden test cases
      description: |
        Maps to `ProblemsApi.submitTests(problemId, code)`. Like `run`, but
        also grades against the problem's hidden test cases. Hidden test
        results are redacted to pass/fail (+ error message) only — their
        input and expected output are never included in the response.
      parameters:
        - $ref: "#/components/parameters/ProblemId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunRequest"
      responses:
        "200":
          description: Grading finished.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmitResponse"
        "404":
          description: No problem with this ID.
...
components:
  parameters:
...
    ProblemId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: The problem's ID, as returned by `GET /api/problems`.
  schemas:
...
    RunRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: Full Python source the candidate has written; must define `functionName`.
    VisibleTestResult:
      type: object
      required:
        - index
        - passed
        - input
        - expected
      properties:
        index:
          type: integer
        passed:
          type: boolean
        input:
          type: array
          items: {}
        expected: {}
        actual:
          description: The candidate function's return value for this test, if it ran without raising.
        error:
          type: string
          nullable: true
          description: Exception message if the candidate function raised instead of returning.
        stdout:
          type: string
          description: Anything the candidate's function printed during this test call (truncated at 4000 chars), for debugging. Empty string if nothing was printed.
    HiddenTestResult:
      type: object
      required:
        - index
        - passed
      description: Same as VisibleTestResult but with input/expected/actual stripped out, so hidden test data never reaches the client.
      properties:
        index:
          type: integer
        passed:
          type: boolean
        error:
          type: string
          nullable: true
    RunResponse:
      type: object
      required:
        - results
        - allPassed
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/VisibleTestResult"
        allPassed:
          type: boolean
        error:
          type: string
          nullable: true
          description: Set instead of `results` when the code couldn't be graded at all (syntax error, missing function, or timeout).
    SubmitResponse:
      type: object
      required:
        - visible
        - hidden
        - passedCount
        - totalCount
        - allPassed
      properties:
        visible:
          type: array
          items:
            $ref: "#/components/schemas/VisibleTestResult"
        hidden:
          type: array
          items:
            $ref: "#/components/schemas/HiddenTestResult"
        passedCount:
          type: integer
        totalCount:
          type: integer
        allPassed:
          type: boolean
        error:
          type: string
          nullable: true